backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
requests
beautifulsoup4
playwright
celery
redis
python-dotenv
openai
google-generativeai
python-multipart
jose[cryptography]
passlib[bcrypt]
